get annual report details

+++ FinancialPlanningBE/FinancialPlanning.Data/Entities/ExpenseAnnualReport.cs
@@ -9,8 +9,9 @@ namespace FinancialPlanning.Data.Entities
     public class ExpenseAnnualReport
     {
         public string Department { get; set; }
-        public int TotalExpense { get; set; }
-        public int BiggestExpenditure { get; set; }
+        public long TotalExpense { get; set; }
+        public long BiggestExpenditure { get; set; }
         public string CostType { get; set; }
+      
     }
 }
